create CommmonCog class
